Move enrollment logic into EnrollmentLogic class
