+++ src/main/java/teammates/logic/core/UsersLogic.java
@@ -4,25 +4,15 @@ import static teammates.common.util.Const.ERROR_UPDATE_NON_EXISTENT;
 
 import java.time.Instant;
 import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
 import java.util.Comparator;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
-import java.util.Set;
-import java.util.StringJoiner;
 import java.util.UUID;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 
 import jakarta.annotation.Nullable;
 
-import teammates.common.datatransfer.EnrollResults;
 import teammates.common.datatransfer.InstructorPermissionRole;
 import teammates.common.datatransfer.InstructorPrivileges;
-import teammates.common.exception.EnrollException;
 import teammates.common.exception.EntityAlreadyExistsException;
 import teammates.common.exception.EntityDoesNotExistException;
 import teammates.common.exception.InstructorUpdateException;
@@ -51,8 +41,6 @@ import teammates.storage.entity.User;
 import teammates.ui.exception.InvalidOperationException;
 import teammates.ui.request.InstructorCreateRequest;
 import teammates.ui.request.InstructorUpdateRequest;
-import teammates.ui.request.StudentEnrollRequest;
-import teammates.ui.request.StudentUpdateRequest;
 
 /**
  * Handles operations related to user (instructor & student).
@@ -61,11 +49,6 @@ import teammates.ui.request.StudentUpdateRequest;
  * @see UsersDb
  */
 public final class UsersLogic {
-
-    static final String ERROR_ENROLL_EXCEED_SECTION_LIMIT =
-            "You are trying enroll more than %s students in section \"%s\".";
-    static final String ERROR_ENROLL_EXCEED_SECTION_LIMIT_INSTRUCTION =
-            "To avoid performance problems, please do not enroll more than %s students in a single section.";
 
     private static final UsersLogic instance = new UsersLogic();
 
@@ -817,10 +800,6 @@ public final class UsersLogic {
         usersDb.deleteStudentsInCourse(courseId);
     }
 
-    private boolean isEmailChanged(String originalEmail, String newEmail) {
-        return newEmail != null && !originalEmail.equals(newEmail);
-    }
-
     /**
      * Updates a student by attributes to update. If an attribute is null, it will not be updated.
      */
@@ -845,202 +824,6 @@ public final class UsersLogic {
         validateUser(student);
 
         return student;
-    }
-
-    /**
-     * Updates a student by student id and update request, and cascades to responses and comments if needed.
-     */
-    public Student updateStudent(UUID studentId, StudentUpdateRequest updateRequest)
-            throws InvalidParametersException, EntityDoesNotExistException, EntityAlreadyExistsException, EnrollException {
-
-        Student student = getStudent(studentId);
-        if (student == null) {
-            throw new EntityDoesNotExistException(String.format("Student with id %s not found", studentId));
-        }
-
-        boolean isEmailChanged = isEmailChanged(student.getEmail(), updateRequest.getEmail());
-        if (isEmailChanged) {
-            Instructor instructor = getInstructorForEmail(student.getCourseId(), updateRequest.getEmail());
-            if (instructor != null) {
-                String errorMessage = String.format(
-                        "Cannot update student email to %s as this email is already used by an instructor in course %s",
-                        updateRequest.getEmail(), student.getCourseId());
-                throw new EntityAlreadyExistsException(errorMessage);
-            }
-
-            Student existingStudent = getStudentForEmail(student.getCourseId(), updateRequest.getEmail());
-            if (existingStudent != null) {
-                String errorMessage = String.format(
-                        "Cannot update student email to %s as this email is already used by another student in course %s",
-                        updateRequest.getEmail(), student.getCourseId());
-                throw new EntityAlreadyExistsException(errorMessage);
-            }
-        }
-
-        // section name -> section
-        Map<String, Section> sections = student.getCourse().getSections()
-                .stream()
-                .collect(Collectors.toMap(Section::getName, Function.identity()));
-
-        // section -> team name -> team
-        Map<String, Map<String, Team>> teams = coursesLogic.getTeamsForCourse(student.getCourseId())
-                .stream()
-                .collect(Collectors.groupingBy(team -> team.getSection().getName(),
-                        Collectors.toMap(Team::getName, Function.identity())));
-
-        Section section = sections.get(updateRequest.getSection());
-        if (section == null) {
-            section = coursesLogic.createSection(student.getCourse(), updateRequest.getSection());
-            sections.put(section.getName(), section);
-        }
-
-        Team team = teams.getOrDefault(section.getName(), Collections.emptyMap()).get(updateRequest.getTeam());
-        if (team == null) {
-            team = coursesLogic.createTeam(section, updateRequest.getTeam());
-            teams.computeIfAbsent(section.getName(), k -> new HashMap<>()).put(team.getName(), team);
-        }
-
-        updateStudentCascade(student, updateRequest.getEmail(), updateRequest.getName(), team, updateRequest.getComments());
-
-        // Validate section limits.
-        // Precondition: this is executed within a transaction; throwing an exception
-        // here will roll back all changes.
-        List<Student> studentsInCourse = getStudentsForCourse(student.getCourseId());
-        String errorMessage = getSectionInvalidityInfo(studentsInCourse);
-        if (!errorMessage.isEmpty()) {
-            throw new EnrollException(errorMessage);
-        }
-
-        return student;
-    }
-
-    /**
-     * Updates a student and enqueues the corresponding feedback session summary email.
-     */
-    public Student updateStudentAndEnqueueSummaryEmail(UUID studentId, StudentUpdateRequest updateRequest)
-            throws InvalidParametersException, EntityDoesNotExistException, EntityAlreadyExistsException, EnrollException {
-        Student student = updateStudent(studentId, updateRequest);
-        feedbackSessionsLogic.enqueueFeedbackSessionSummaryEmail(student, EmailType.STUDENT_EMAIL_CHANGED);
-        return student;
-    }
-
-    /**
-     * Enrolls students in a course according to the enroll requests, creating the section and team if needed.
-     */
-    public EnrollResults enrollStudents(Course course,
-            List<StudentEnrollRequest> enrollRequests) throws EnrollException {
-        Set<String> instructorEmails = getInstructorsForCourse(course.getId())
-                .stream()
-                .map(Instructor::getEmail)
-                .collect(Collectors.toSet());
-
-        // student email -> student
-        Map<String, Student> studentsInCourse = getStudentsForCourse(course.getId())
-                .stream()
-                .collect(Collectors.toMap(Student::getEmail, Function.identity()));
-
-        // section name -> section
-        Map<String, Section> sections = course.getSections()
-                .stream()
-                .collect(Collectors.toMap(Section::getName, Function.identity()));
-
-        // section -> team name -> team
-        Map<String, Map<String, Team>> teams = coursesLogic.getTeamsForCourse(course.getId())
-                .stream()
-                .collect(Collectors.groupingBy(team -> team.getSection().getName(),
-                        Collectors.toMap(Team::getName, Function.identity())));
-
-        EnrollResults enrollResults = new EnrollResults();
-
-        // Process individual enroll requests
-        for (StudentEnrollRequest enrollRequest : enrollRequests) {
-            String email = enrollRequest.getEmail();
-            if (instructorEmails.contains(email)) {
-                String errorMsg = String.format(
-                        "Cannot enroll student with email %s as this email is already used by an instructor in course %s",
-                        email, course.getId());
-                enrollResults.addUnsuccessfulEnroll(email, errorMsg);
-                continue;
-            }
-
-            try {
-                Student student = processEnrollRequest(course, studentsInCourse, sections, teams, enrollRequest);
-                enrollResults.addEnrolledStudent(student);
-            } catch (InvalidParametersException | EntityAlreadyExistsException e) {
-                enrollResults.addUnsuccessfulEnroll(email, e.getMessage());
-            }
-        }
-
-        // Validate section limit violations.
-        // Precondition: this is executed within a transaction; throwing an exception here will roll back all changes.
-        String errorMessage = getSectionInvalidityInfo(studentsInCourse.values());
-        if (!errorMessage.isEmpty()) {
-            throw new EnrollException(errorMessage);
-        }
-
-        return enrollResults;
-    }
-
-    /**
-     * Process an individual enroll request, creating the section, team and student if needed.
-     */
-    private Student processEnrollRequest(
-            Course course,
-            Map<String, Student> studentsInCourse,
-            Map<String, Section> sections,
-            Map<String, Map<String, Team>> teams,
-            StudentEnrollRequest enrollRequest) throws InvalidParametersException, EntityAlreadyExistsException {
-        String email = enrollRequest.getEmail();
-
-        Section section = sections.get(enrollRequest.getSection());
-        if (section == null) {
-            section = coursesLogic.createSection(course, enrollRequest.getSection());
-            sections.put(section.getName(), section);
-        }
-
-        Team team = teams.getOrDefault(section.getName(), Collections.emptyMap()).get(enrollRequest.getTeam());
-        if (team == null) {
-            team = coursesLogic.createTeam(section, enrollRequest.getTeam());
-            teams.computeIfAbsent(section.getName(), k -> new HashMap<>()).put(team.getName(), team);
-        }
-
-        Student student = studentsInCourse.get(email);
-        if (student != null) {
-            updateStudentCascade(student, null, enrollRequest.getName(), team, enrollRequest.getComments());
-        } else {
-            student = createStudent(course, team, enrollRequest.getName(), email, enrollRequest.getComments());
-            studentsInCourse.put(email, student);
-        }
-
-        return student;
-    }
-
-    private String getSectionInvalidityInfo(Collection<Student> studentList) {
-        Map<String, Integer> sectionCountMap = new HashMap<>();
-        for (Student student : studentList) {
-            String sectionName = student.getSectionName();
-            assert sectionName != null : "Section name should not be null";
-            sectionCountMap.put(sectionName, sectionCountMap.getOrDefault(sectionName, 0) + 1);
-        }
-
-        StringJoiner errorMessage = new StringJoiner(" ");
-        sectionCountMap.entrySet().stream()
-                .sorted(Map.Entry.comparingByKey())
-                .forEach(entry -> {
-                    if (entry.getValue() > Const.SECTION_SIZE_LIMIT) {
-                        errorMessage.add(String.format(
-                                ERROR_ENROLL_EXCEED_SECTION_LIMIT,
-                                Const.SECTION_SIZE_LIMIT, entry.getKey()));
-                    }
-                });
-
-        if (errorMessage.length() > 0) {
-            errorMessage.add(String.format(
-                    ERROR_ENROLL_EXCEED_SECTION_LIMIT_INSTRUCTION,
-                    Const.SECTION_SIZE_LIMIT));
-        }
-
-        return errorMessage.toString();
     }
 
     /**
