+++ src/main/java/teammates/ui/webapi/UpdateStudentAction.java
@@ -7,7 +7,6 @@ import teammates.common.exception.EntityAlreadyExistsException;
 import teammates.common.exception.EntityDoesNotExistException;
 import teammates.common.exception.InvalidParametersException;
 import teammates.common.util.Const;
-import teammates.storage.entity.Student;
 import teammates.ui.exception.EntityNotFoundException;
 import teammates.ui.exception.InvalidHttpRequestBodyException;
 import teammates.ui.exception.InvalidOperationException;
@@ -18,32 +17,10 @@ import teammates.ui.request.StudentUpdateRequest;
  * Action: Edits details of a student in a course.
  */
 public class UpdateStudentAction extends LoggedInAction {
-    /** Message indicating that the student to be edited could not be found in the system. */
-    public static final String STUDENT_NOT_FOUND_FOR_EDIT = "The student you tried to edit does not exist.";
-
-    /** Message indicating that the student information was successfully updated. */
-    public static final String SUCCESSFUL_UPDATE = "Student has been updated";
-    /**
-     * Message indicating that the student information was successfully updated,
-     * and email was queued to the student's new email address.
-     */
-    public static final String SUCCESSFUL_UPDATE_WITH_EMAIL = SUCCESSFUL_UPDATE + " and email sent";
-    /**
-     * Message indicating that the update operation failed because the requested new email address
-     * for the student is already being used by another student in the system.
-     */
-    public static final String ERROR_EMAIL_ALREADY_EXISTS = "Trying to update to an email that is already in use";
-
     @Override
     void checkSpecificAccessControl() throws UnauthorizedAccessException {
         UUID studentId = getUuidRequestParamValue(Const.ParamsNames.USER_ID);
-        Student existingStudent = logic.getStudent(studentId);
-        if (existingStudent == null) {
-            throw new EntityNotFoundException(STUDENT_NOT_FOUND_FOR_EDIT);
-        }
-
-        gateKeeper.verifyInstructorHasPrivilege(requestContext, existingStudent.getCourseId(),
-                Const.InstructorPermissions.CAN_MODIFY_STUDENT);
+        gateKeeper.verifyCanModifyStudent(requestContext, studentId);
     }
 
     @Override
@@ -51,27 +28,19 @@ public class UpdateStudentAction extends LoggedInAction {
         UUID studentId = getUuidRequestParamValue(Const.ParamsNames.USER_ID);
         StudentUpdateRequest updateRequest = getAndValidateRequestBody(StudentUpdateRequest.class);
 
-        Student existingStudent = logic.getStudent(studentId);
-        if (existingStudent == null) {
-            throw new EntityNotFoundException(STUDENT_NOT_FOUND_FOR_EDIT);
-        }
-
         try {
             if (updateRequest.getIsSessionSummarySendEmail()) {
                 logic.updateStudentAndEnqueueSummaryEmail(studentId, updateRequest);
-                return new JsonResult(SUCCESSFUL_UPDATE_WITH_EMAIL);
             } else {
-                logic.updateStudent(studentId, updateRequest);
-                return new JsonResult(SUCCESSFUL_UPDATE);
+                logic.updateStudentEnrollment(studentId, updateRequest);
             }
-        } catch (EnrollException e) {
+            return new JsonResult("Student update successfully");
+        } catch (EnrollException | EntityAlreadyExistsException e) {
             throw new InvalidOperationException(e);
         } catch (InvalidParametersException e) {
             throw new InvalidHttpRequestBodyException(e);
         } catch (EntityDoesNotExistException ednee) {
             throw new EntityNotFoundException(ednee);
-        } catch (EntityAlreadyExistsException e) {
-            throw new InvalidOperationException(ERROR_EMAIL_ALREADY_EXISTS, e);
         }
     }
